CHROMA_Synthetica_v1.0/requirements.txt
# Nenhuma dependência externa necessária por enquanto.
# Adicionaremos aqui se usarmos APIs ou outras bibliotecas.
google-generativeai>=0.5.0
